requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
requests
python-dotenv
azure-eventhub
requests
databricks-vectorsearch
mlflow
databricks-sdk